+++ src/main/java/com/groom/orbit/auth/dao/AuthMemberRepository.java
@@ -1,0 +1,12 @@
+package com.groom.orbit.auth.dao;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.groom.orbit.auth.dao.entity.AuthMember;
+
+public interface AuthMemberRepository extends JpaRepository<AuthMember, Long> {
+
+  Optional<AuthMember> findByNickname(String email);
+}
