+++ src/main/java/com/groom/orbit/auth/dao/entity/AuthMember.java
@@ -6,13 +6,14 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.ColumnDefault;
 
 @Entity(name = "member")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Member {
+public class AuthMember {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -22,23 +23,8 @@ public class Member {
   @Column(name = "nickname", length = 100)
   private String nickname;
 
+  @ColumnDefault("")
   @Column(name = "image_url")
   @Builder.Default
   private String imageUrl = "";
-
-  @Column(name = "known_prompt", length = 1000)
-  @Builder.Default
-  private String knownPrompt = "";
-
-  @Column(name = "help_prompt", length = 1000)
-  @Builder.Default
-  private String helpPrompt = "";
-
-  @Column(name = "is_notification")
-  @Builder.Default
-  private Boolean isNotification = false;
-
-  @Column(name = "is_profile")
-  @Builder.Default
-  private Boolean isProfile = false;
 }
